cpp. aggregated real scans. needs logic check.

+++ cpp/include/lidar_sim/PoseServer.h
@@ -12,7 +12,7 @@ namespace lidar_sim {
 	void printLogAtIndex(int data_id);
 	void printPose(const std::vector<double> pose);
 	std::vector<double> getPoseAtTime (double t) const;
-	Eigen::Matrix<float,4,4> getTransfAtTime(double t);
+	Eigen::Matrix<double,4,4> getTransfAtTime(double t);
 	std::tuple<std::vector<double>, int> 
 	    getNearestPoseInLog(const std::vector<double> pose);
